# openapi.yaml
openapi: 3.0.3
info:
  title: SEBT_platform API
  version: 1.3.1
  description: >-
    Scope simplified: removed escrow & withdrawal (out-of-scope). Added listing
    publish endpoint with listing fee debit. Wallet supports top-up and listing
    fee only. Favorites legacy endpoints remain deprecated.
servers:
  - url: http://localhost:8080/
    description: Local dev
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  schemas:
...
    WalletBalanceResponse:
      type: object
      properties:
        balance: { type: number, format: double }
      required: [balance]
    WalletTransaction:
      type: object
      properties:
        orderId: { type: string }
        amount: { type: number, format: double }
        balanceBefore: { type: number, format: double }
        balanceAfter: { type: number, format: double }
        status: { type: string }
        purpose: { type: string }
        entryType: { type: string }
        createdAt: { type: string, format: date-time }
        description: { type: string }
      required: [orderId, amount, status]
...
paths:
...
  /wallet/me:
    get:
      summary: Lấy số dư ví hiện tại
      tags: [Wallet]
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Wallet balance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletBalanceResponse' }
        '401': { description: Unauthorized }
  /wallet/transactions:
    get:
      summary: Lịch sử giao dịch ví
      tags: [Wallet]
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 10 } }
        - { name: purpose, in: query, schema: { type: string } }
      responses:
        '200':
          description: Page of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items: { $ref: '#/components/schemas/WalletTransaction' }
                  totalElements: { type: integer }
                  totalPages: { type: integer }
                  size: { type: integer }
                  number: { type: integer }
        '401': { description: Unauthorized }
...